// src/pages/device/sections/DeviceDetails/_forms/deviceEditForm.types.ts
export interface FormData {
  machine_Name: string;
  waste: string;
  status: number;
  linuxVersion: string;
  startAvailable: string;
  endAvailable: string;
  street: string;
  postalCode: string;
  province: string;
  city: string;
  country: string;
  municipality: string;
  address: string;
  tatus_ready_d75_3_7: boolean;
  status_Machine_Blocked: boolean;
  codiceGps: string;
  sheetName: string;
  customerName: string;
  matricolaBte: string;
  matricolaKgn: string;
  customer: string;
  ip_Router: string;
  gpsX: string;
  gpsY: string;
  note: string;
}